vect.h: fix severe index calculation bug
C will not automatically multiply our pointer by the size of our type
when using uintptr_t. This caused misaligned accesses and invalid
results to be returned for any case other than single byte types or the
first element in the vector (which happened to be the only test cases we
had!).